+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Added
 - Label: Can render with a hidden field - set `include_hidden_field: true`.
 - Label: Optionally provide a different value for the hidden field using `:hidden_value`.
+- Grids: Can pass in columns and rows directly so the grid loads with the page instead of using a URL to load rows via a delayed request.
+- Nodes can add javascript to be included in the page. Currently only used by Grid.
 ### Changed
 ### Fixed